requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & JWT
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0

# Image handling
Pillow==10.2.0

# File storage (AWS S3)
boto3==1.34.34
django-storages==1.14.2

# Environment variables
python-dotenv==1.0.1

# Filtering and Search
django-filter==23.5
django-haystack==3.2.1
elasticsearch==8.12.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Background tasks
celery==5.3.6
celery[redis]==5.3.6

# API Documentation
drf-spectacular==0.27.1

# Phone number validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.29

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Security & Monitoring
django-ratelimit==4.1.0
sentry-sdk==1.40.0

# Utilities
python-slugify==8.0.2
bleach==6.1.0

# Real-time messaging (optional)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# SMS (Twilio)
twilio==8.13.0

# Gunicorn for production
gunicorn==21.2.0

# Whitenoise for static files
whitenoise==6.6.0